[Bugfix] Disable w16a16 2of4 sparse CompressedTensors24 (#12417)
